refact: configure BigDecimal in values

+++ src/main/kotlin/br/com/kotlin/learning/model/Rent.kt
@@ -1,5 +1,6 @@
 package br.com.kotlin.learning.model
 
+import java.math.BigDecimal
 import java.time.LocalDate
 import java.time.Period
 
@@ -8,7 +9,7 @@ data class Rent(
     val game: Game,
     val rentalPeriod: RentalPeriod
 ){
-    private val rentValue: Double = gamerUser.plan.getPrice(this)
+    private val rentValue: BigDecimal = gamerUser.plan.getPrice(this)
     override fun toString(): String {
         return "${gamerUser.name} rent the game ${game.title} " +
                 "per R$$rentValue (R$${game.price}/day) " +
